+++ app/serializers/api/v1/project_serializer.rb
@@ -10,7 +10,7 @@ module Api
         markdown(object.description)
       end
 
-      link(:self) { api_v1_project_path(object.id) }
+      link(:self) { api_project_path(object.id) }
       link(:github) { "https://github.com/#{object.github_url}" }
       has_one :project_status, key: :status
       has_many :project_categories, key: :categories
